rename LineupLoader to Lineup, rename ImageLoader to Photo, rename ImageAnalyzer to PhotoAnalyzer.
