+++ src/pyWitnessAI/Photo.py
